autocmd bufwritepost .vimrc source $MYVIMRC

+++ setup.sh
@@ -5,11 +5,28 @@ git clone git://github.com/robbyrussell/oh-my-zsh.git ~/.oh-my-zsh
 
 chsh -s /bin/zsh
 
-ln -s ~/.dotfiles/vim/vimrc ~/.vimrc
-ln -s ~/.dotfiles/vim ~/.vim
+if [ -e ~/.vimrc ]; then
+    echo "vimrc already exist"
+else
+    ln -s ~/.dotfiles/vim/vimrc ~/.vimrc
+fi
 
-ln -s ~/.dotfiles/zsh/zshrc ~/.zshrc
+if [ -e ~/.vim ]; then
+    echo "vim folder already exist"
+else
+    ln -s ~/.dotfiles/vim ~/.vim
+fi
 
-ln -s ~/.dotfiles/tmux/tmux.conf ~/.tmux.conf
+if [ -e ~/.zshrc ]; then
+    echo "zshrc already exist"
+else
+    ln -s ~/.dotfiles/zsh/zshrc ~/.zshrc
+fi
+
+if [ -e ~/.zshrc ]; then
+    echo "tmux.conf already exist"
+else
+    ln -s ~/.dotfiles/tmux/tmux.conf ~/.tmux.conf
+fi
 
 echo "Please BundleInstall~~~~~~~~~~~~~~~~~"
